rag-backend/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alibabacloud-oss-v2>=1.1.3",
    "docx>=0.2.4",
    "fastapi>=0.116.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "langchain>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-milvus>=0.2.1",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langgraph-cli[inmem]>=0.4.2",
    "psycopg[binary,pool]>=3.0.0",
    "psycopg2-binary>=2.9.9",
    "pymilvus>=2.6.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
    "pymysql>=1.1.2",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langchain-qwq>=0.2.1",
    "langchain-deepseek>=0.1.4",
    "langmem>=0.0.29",
    "langsmith>=0.4.25",
    "langchain-mcp-adapters>=0.1.10",
    "lightrag-hku>=1.4.8.2",
    "bcrypt>=4.3.0",
    "asyncpg>=0.30.0",
    "neo4j>=5.28.2",
    "crawl4ai>=0.7.4",
    "redis>=5.0.0",
    "pytz>=2024.1",
    "cos-python-sdk-v5>=1.9.40",
    "mcp>=1.15.0",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
extra-index-url = [
    "https://mirrors.aliyun.com/pypi/simple",
    "https://mirrors.ustc.edu.cn/pypi/simple"
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-mock>=3.15.0",
]
